+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - DEB: logrotate config
 - API: more exports to metrics
+- API: config endpoint
 
 ### Fixed
 - API: return change lists (stations, checks, clicks) from start if lastuuid not found